move `migrate-drop-unnecessary-data-types` to core
